Create symbol.yaml
One of the tricky things is what to do with the symbol. Essentially a weather type icon is 2 numbers and  either d for day or n for night and there are a lot of them. 
The problem is you can't just make a url and give it as the image path.  what you can do is give it a defaut image and a number of state images which really is pain as there are a lot of possibilities if you see the default image the weather report has almost certainly used a state you haven't catered for.